+++ specification/ml/put_filter/MlPutFilterResponse.ts
@@ -20,5 +20,9 @@
 import { Id } from '@_types/common'
 
 export class Response {
-  body: { description: string; filter_id: Id; items: string[] }
+  body: {
+    description: string
+    filter_id: Id
+    items: string[]
+  }
 }
