+++ specification/ml/put_job/MlPutJobResponse.ts
@@ -36,7 +36,7 @@ export class Response {
     job_id: Id
     job_type: string
     model_plot: ModelPlotConfig
-    model_snapshot_id: string
+    model_snapshot_id: Id
     model_snapshot_retention_days: long
     renormalization_window_days: long
     results_index_name: string
